refactor(landing): 首页拆分为独立 Vue 组件
